new way of config

+++ src/DeleteMethod.cpp
@@ -45,8 +45,8 @@ std::string DeleteMethod::generateResponse(Request req, ServerConfigs serv)
 					std::string path = "/" + trim(mygetcwd(), '/') + "/" + trim(loc.root, '/') + "/" + trim(req._path, '/');
 					if (access(path.c_str(), F_OK) != 0)
 						return err404(req._version); // se non esiste
-					// if (access(path.c_str(), W_OK) != 0 || access(path.c_str(), X_OK) != 0 || access(path.c_str(), R_OK) != 0)
-					// 	return err403(req._version); // se non hai permessi devi avere tutti e 3 o solo scrittura e lettura non si sa
+					if (access(path.c_str(), W_OK) != 0 || access(path.c_str(), X_OK) != 0 || access(path.c_str(), R_OK) != 0)
+						return err403(req._version); // se non hai permessi devi avere tutti e 3 o solo scrittura e lettura non si sa
 					if (removeFile(path.c_str()) != 0)
 						return err500(req._version); // se remove fallisce perchè il kernel è stronzo
 					response += req._version + " 200 OK\r\n";
@@ -58,23 +58,7 @@ std::string DeleteMethod::generateResponse(Request req, ServerConfigs serv)
 			return err405(req._version); // metodo non permesso
 		}
 		else
-		{
-			if (checkfile(trim(temp.root, '/') + "/" + req._path, tmp) == 0)
-				return err403(req._version);
-			if (checkfile(trim(temp.root, '/') + "/" + req._path, tmp) == 1)
-				filename = tmp;
-			std::string path = "/" + trim(mygetcwd(), '/') + "/" + trim(temp.root, '/') + "/" + trim(req._path, '/');
-			if (access(path.c_str(), F_OK) != 0)
-				return err404(req._version); // se non esiste
-			// if (access(path.c_str(), W_OK) != 0 || access(path.c_str(), X_OK) != 0 || access(path.c_str(), R_OK) != 0)
-			// 	return err403(req._version); // se non hai permessi devi avere tutti e 3 o solo scrittura e lettura non si sa
-			if (removeFile(path.c_str()) != 0)
-				return err500(req._version); // se remove fallisce perchè il kernel è stronzo
-			response += req._version + " 200 OK\r\n";
-			response += "Content-Type: text/html\r\n\r\n";
-			response += getFile("/" + trim(mygetcwd(),'/') + "/" + trim(temp.root, '/') + "/" + "delete.html");
-			return response;
-		}
+			return err404(req._version);
 	}
 	return err400(req._version); // non trova il config per qualche motivo a me sconosciuto senno da errore in compilazione
 }
